+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Development
 
+* Add `Node#ancestors`, `Node#child_nodes` and `Node#descendants` which return an array instead of enumerator.
+
 ## v1.0.0
 
 * Initial API stable release in the Semantic Versioning.